websockets criteria

+++ tests/testsuite/src/main/java/de/quantummaid/httpmaid/tests/givenwhenthen/builders/FirstWhenStage.java
@@ -41,7 +41,7 @@ public interface FirstWhenStage {
 
     Then theLastWebsocketIsDisconnected();
 
-    Then theRuntimeDataIsQueriedUntilTheNumberOfWebsocketsBecomes(int expectedNumberOfWebsockets);
+    Then theRuntimeDataIsQueriedUntilTheNumberOfWebsocketsBecomes(long expectedNumberOfWebsockets);
 
     Then theRuntimeDataIsQueried();
 }
